feat: implement start game

+++ src/ws_server/controllers/index.ts
@@ -1,29 +1,30 @@
-import { UserDB, wsAPI, wsMsg } from '../types';
-import { generateID, stringifyData } from '../utils';
-import { User, getWinners } from './usersController';
+import { RoomDB, UserDB, UserFront, wsAPI, wsMsg } from '../types';
+import { fontLog, generateID, stringifyData } from '../utils';
+import { User, getWinners, userExists, usersDB } from './usersController';
 import { IncomingMessage } from 'http';
 import { wss } from '..';
-import { Room, getRoomById, roomsDB } from './roomsController';
+import { Room, getRoomById, roomExists, roomsDB } from './roomsController';
+import { Game, gameExists, gamesDB, getGameById } from './gamesController';
 
 const CLIENTS = {};
 export const handleWS = (ws: WebSocket, req: IncomingMessage) => {
 	const id = generateID();
-	console.log('ID: ', id);
 	CLIENTS[id] = ws;
 
 	let curUser: UserDB;
 
 	const broadcast = (msg: string, ids = []) => {
-		console.log('->> Broadcast:', msg);
+		console.log(fontLog.FgCyan, '->> Broadcast:', msg);
 		if (ids.length) {
+			console.log(ids);
 			ids.forEach(id => {
 				const targetClient = CLIENTS[id];
 				targetClient.send(msg);
 			})
 		} else {
 			wss.clients.forEach((client) => {
-					client.send(msg);
-			 });
+				client.send(msg);
+			});
 		}
 	};
 
@@ -37,9 +38,10 @@ export const handleWS = (ws: WebSocket, req: IncomingMessage) => {
 		
 		switch (type) {
 			case wsAPI.reg:
-				let user = new User({...data});
+				const userInput: UserFront = data;
+				let user = new User({...userInput});
 				
-				if (user.exists() && !user.validPass()) {
+				if (userExists(userInput.name) && !user.validPass()) {
 					user.error = true;
 					user.errorText = 'User already exists, wrong password';
 				} else {
@@ -54,33 +56,49 @@ export const handleWS = (ws: WebSocket, req: IncomingMessage) => {
 
 			case wsAPI.createRoom:
 				const room = new Room(curUser);
-				if (!room.exists(curUser.name)) {
+				if (!roomExists(curUser.name)) {
 					room.create();
 					// broadcast to all clients about new room creation
 					broadcast(stringifyData(wsAPI.updateRoom, roomsDB));
 				} else {
-					console.log('This user already has one room')
+					console.log(fontLog.FgRed, 'This user already has one room')
 				}
 				break;
 
+			// join somebody's room
 			case wsAPI.joinRoom:
 				const targetRoom = getRoomById(data.indexRoom);
-				targetRoom.addUser(curUser);
-				targetRoom.delete();
+				if (!targetRoom.isMyOwn(curUser)) {
+					targetRoom.addUser(curUser);
+					// broadcast for TWO users - game creation
+					const roomUsersIds = targetRoom.roomUsers.map((u: UserDB)  => u.index);
+					broadcast(stringifyData(wsAPI.createGame, {
+						idGame: targetRoom.roomId,
+						idPlayer: curUser.index,
+					}), roomUsersIds);
 
-				// broadcast ALL: remove room from available
-				broadcast(stringifyData(wsAPI.updateRoom, roomsDB));
-
-				// broadcast TWO users for game creation
-				broadcast(stringifyData(wsAPI.createGame, {
-					idGame: targetRoom.roomId,
-					idPlayer: curUser.index,
-				}), []);
-
+					targetRoom.delete();
+					// broadcast ALL: remove room from available
+					broadcast(stringifyData(wsAPI.updateRoom, roomsDB));
+				} else {
+					console.log(fontLog.FgRed, 'This your room, try to find a real opponent');
+				}
 				break;	
 			
 			case wsAPI.addShips:
+				if (gameExists(data.gameId)) {
+					const game = getGameById(data.gameId);
+					game.addOpponent(data);
 
+					ws.send(stringifyData(wsAPI.startGame, curUser));
+
+					// start game for each user
+					broadcast(stringifyData(wsAPI.startGame, game.player1), [game.player1.currentPlayerIndex]);
+					broadcast(stringifyData(wsAPI.startGame, game.player2), [game.player2.currentPlayerIndex]);
+				} else {
+					const game = new Game();
+					game.create(data);
+				}
 				break;
 		}
 	};
